Update views.py
PUT request is functional. It's also been tested for an unauthorised user trying to update an exisiting workshop details.

+++ pancakes/workshops/views.py
@@ -55,9 +55,22 @@ class WorkshopDetail(APIView):
         serializer = WorkshopDetailSerializer(workshop)
         return Response(serializer.data)
 
+    #PUT request
+    def put(self, request, pk):
+        workshop = self.get_object(pk)
+        if workshop.organiser!=request.user:
+            return HttpResponse(status=403)
+        serializer = WorkshopDetailSerializer(workshop, data=request.data, partial=True)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST) 
+
     # DEL request
     def delete(self, request, pk):
         workshop = self.get_object(pk)
+        if workshop.organiser!=request.user:
+            return HttpResponse(status=403)
         if workshop: 
             workshop.delete()
-        return HttpResponse(status=200)
+            return HttpResponse(status=200)
